Move getRawComments into its own file
